requirements.txt

torch>=2.1
flash-attn>=2.3.0
triton>=2.1.0
pyyaml
tqdm
